+++ CHANGELOG.md
@@ -12,6 +12,9 @@
 
 ### Added
 
+- The `max_line_length` key is now supported. From the Helix documentation:
+  > Used for the `:reflow` command and soft-wrapping if `soft-wrap.wrap-at-text-width` is set
+
 ### Fixed
 
 ## 1.2.0 - 2025-02-07